+++ lib/easyrtc_default_event_listeners.js
@@ -8,13 +8,13 @@
  * @license     BSD v2, see LICENSE file in module root folder.
  */
 
-var util        = require("util");                  // General utility functions core module
-var _           = require("underscore");            // General utility functions external module
-var g           = require("./general_util");        // General utility functions local module
+// Externals dependencies
+var util = require("util"); // General utility functions core module
+var _ = require("underscore"); // General utility functions external module
+var async = require("async"); // Asynchronous calls external module
 
-var async       = require("async");                 // Asynchronous calls external module
-
-var pub         = require("./easyrtc_public_obj");  // EasyRTC public object
+// Internals dependencies
+var pub = require("./easyrtc_public_obj"); // EasyRTC public object
 
 /**
  * Event listeners used by EasyRTC. Many of these can be overridden using server options. The interfaces should be used as a guide for creating new listeners.
@@ -1757,7 +1757,7 @@ eventListener.onStartup = function(next) {
             pub.util.logDebug("Configuring Socket server");
 
             pub.socketServer.sockets.on("connection", function (socket) {
-                var easyrtcid = pub.util.getAvailableEasyrtcid();
+                var easyrtcid = pub.getAvailableEasyrtcid();
 
                 pub.util.logDebug("["+easyrtcid+"]["+socket.id+"] Socket connected");
                 pub.util.logDebug("Emitting event 'connection'");
