+++ lib/easyrtc_private_obj.js
@@ -8,15 +8,15 @@
  * @license     BSD v2, see LICENSE file in module root folder.
  */
 
-// var _               = require("underscore");                // General utility functions external module
-var defaultOptions  = require("./easyrtc_default_options"); // EasyRTC global variable
-var g               = require("./general_util");            // General utility functions local module
+// Internals dependencies
+var defaultOptions = require("./easyrtc_default_options");    // EasyRTC global variable
+var utils = require("./utils");                      // General utility functions local module
 
-var e = {};
+var easyrtc = {};
 
-e.version           = g.getPackageData("version");
-e.serverStartOn     = Date.now();
-e.option            = g.deepCopy(defaultOptions);
-e.app               = {};
+easyrtc.version           = utils.getPackageData("version");
+easyrtc.serverStartOn     = Date.now();
+easyrtc.option            = utils.deepCopy(defaultOptions);
+easyrtc.app               = {};
 
-module.exports = e;
+module.exports = easyrtc;
