+++ lib/easyrtc_server.js
@@ -7,12 +7,15 @@
  * @license     BSD v2, see LICENSE file in module root folder.
  */
 
-var g           = require("./general_util");        // General utility functions local module
-g.checkModules(); // Check to ensure all required modules are available
+// Externals dependencies
+var _ = require("underscore"); // General utility functions external module
 
-var _           = require("underscore");            // General utility functions external module
-var pub         = require("./easyrtc_public_obj");  // EasyRTC public object
+// Internals dependencies
+var pub = require("./easyrtc_public_obj"); // EasyRTC public object
+var utils = require("./utils"); // General utility functions local module
+var easyrtc_utils = require("./easyrtc_utils"); // General utility functions local module
 
+utils.checkModules(); // Check to ensure all required modules are available
 
 var server = module.exports;
 
@@ -66,7 +69,7 @@ server.listen = function(httpApp, socketServer, options, listenCallback) {
  */
 server.getDefaultOptions = function() {
     var defaultOptions = require("./easyrtc_default_options");
-    return g.deepCopy(defaultOptions);
+    return utils.deepCopy(defaultOptions);
 };
 
 /**
